fix: OpenSSL version is downgraded to working one

+++ summa-core/src/components/fruit_extractors.rs
@@ -141,7 +141,7 @@ pub fn build_fruit_extractor(
                     scorer: Some(proto::scorer::Scorer::OrderBy(field_name)),
                 }) => {
                     let top_docs_collector = tantivy::collector::TopDocs::with_limit((top_docs_collector_proto.limit + 1) as usize)
-                        .and_offset(top_docs_collector_proto.offset)
+                        .and_offset(top_docs_collector_proto.offset as usize)
                         .order_by_fast_field(field_name, Order::Desc);
                     Box::<TopDocs<u64>>::new(
                         TopDocsBuilder::default()
